+++ src/Element/Script.php
@@ -4,112 +4,40 @@ declare(strict_types = 1);
 namespace Innmind\Html\Element;
 
 use Innmind\Xml\{
-    Node,
-    Attribute,
     Element,
-    Node\Text,
+    Element\Name,
+    Element\Custom,
+    Attribute,
+    Node,
 };
-use Innmind\Immutable\{
-    Set,
-    Sequence,
-    Maybe,
-    Map,
-};
+use Innmind\Immutable\Sequence;
 
 /**
  * @psalm-immutable
  */
-final class Script implements Element
+final class Script implements Custom
 {
-    private Element\Element $element;
-
-    private function __construct(Element\Element $element)
+    private function __construct(private Element $element)
     {
-        $this->element = $element;
     }
 
     /**
      * @psalm-pure
      *
-     * @param Set<Attribute>|null $attributes
+     * @param Sequence<Attribute>|null $attributes
      */
-    public static function of(Text $text, ?Set $attributes = null): self
+    public static function of(Node $text, ?Sequence $attributes = null): self
     {
-        /** @var Sequence<Node> */
-        $children = Sequence::of($text);
-
-        return new self(Element\Element::of('script', $attributes, $children));
+        return new self(Element::of(
+            Name::of('script'),
+            $attributes,
+            Sequence::of($text),
+        ));
     }
 
     #[\Override]
-    public function name(): string
+    public function normalize(): Element
     {
-        return 'script';
-    }
-
-    #[\Override]
-    public function attributes(): Map
-    {
-        return $this->element->attributes();
-    }
-
-    #[\Override]
-    public function attribute(string $name): Maybe
-    {
-        return $this->element->attribute($name);
-    }
-
-    #[\Override]
-    public function removeAttribute(string $name): self
-    {
-        return new self($this->element->removeAttribute($name));
-    }
-
-    #[\Override]
-    public function addAttribute(Attribute $attribute): self
-    {
-        return new self($this->element->addAttribute($attribute));
-    }
-
-    #[\Override]
-    public function children(): Sequence
-    {
-        return $this->element->children();
-    }
-
-    #[\Override]
-    public function filterChild(callable $filter): self
-    {
-        return new self($this->element->filterChild($filter));
-    }
-
-    #[\Override]
-    public function mapChild(callable $map): self
-    {
-        return new self($this->element->mapChild($map));
-    }
-
-    #[\Override]
-    public function prependChild(Node $child): self
-    {
-        return new self($this->element->prependChild($child));
-    }
-
-    #[\Override]
-    public function appendChild(Node $child): self
-    {
-        return new self($this->element->appendChild($child));
-    }
-
-    #[\Override]
-    public function content(): string
-    {
-        return $this->element->content();
-    }
-
-    #[\Override]
-    public function toString(): string
-    {
-        return $this->element->toString();
+        return $this->element;
     }
 }