+++ src/Visitor/Element.php
@@ -3,11 +3,17 @@ declare(strict_types = 1);
 
 namespace Innmind\Html\Visitor;
 
+use Innmind\Html\Node\Document;
 use Innmind\Xml\{
     Node,
-    Element as ElementInterface,
+    Element as Model,
+    Element\Custom,
 };
-use Innmind\Immutable\Maybe;
+use Innmind\Immutable\{
+    Maybe,
+    Sequence,
+    Predicate\Instance,
+};
 
 /**
  * @psalm-immutable
@@ -26,24 +32,16 @@ final class Element
     }
 
     /**
-     * @return Maybe<ElementInterface>
+     * @return Maybe<Model|Custom>
      */
-    public function __invoke(Node $node): Maybe
+    public function __invoke(Document|Node|Model|Custom $node): Maybe
     {
-        if (
-            $node instanceof ElementInterface &&
-            $node->name() === $this->name
-        ) {
-            return Maybe::just($node);
-        }
-
-        /** @var Maybe<ElementInterface> */
-        return $node->children()->reduce(
-            Maybe::nothing(),
-            function(Maybe $element, Node $child): Maybe {
-                return $element->otherwise(fn() => $this($child));
-            },
-        );
+        return match (true) {
+            $node instanceof Document => $this->visitDocument($node),
+            $node instanceof Node => $this->visitNode($node),
+            $node instanceof Model => $this->visitElement($node),
+            $node instanceof Custom => $this->visitCustom($node),
+        };
     }
 
     /**
@@ -70,5 +68,74 @@ final class Element
     public static function body(): self
     {
         return new self('body');
+    }
+
+    /**
+     * @return Maybe<Model|Custom>
+     */
+    private function visitDocument(Document $document): Maybe
+    {
+        return $this->visit($document->children());
+    }
+
+    /**
+     * @return Maybe<Model|Custom>
+     */
+    private function visitNode(Node $node): Maybe
+    {
+        /** @var Maybe<Model|Custom> */
+        return Maybe::nothing();
+    }
+
+    /**
+     * @return Maybe<Model|Custom>
+     */
+    private function visitElement(Model $element): Maybe
+    {
+        if ($element->name()->toString() === $this->name) {
+            return Maybe::just($element);
+        }
+
+        return $this->visit($element->children());
+    }
+
+    /**
+     * @return Maybe<Model|Custom>
+     */
+    private function visitCustom(Custom $element): Maybe
+    {
+        $normalized = $element->normalize();
+
+        if ($normalized->name()->toString() === $this->name) {
+            return Maybe::just($element);
+        }
+
+        return $this->visit($normalized->children());
+    }
+
+    /**
+     * @param Sequence<Model|Custom|Node> $children
+     *
+     * @return Maybe<Model|Custom>
+     */
+    private function visit(Sequence $children): Maybe
+    {
+        /** @var Model|Custom|null */
+        $found = null;
+
+        $found = $children
+            ->sink($found)
+            ->until(
+                fn($found, $child, $continuation) => $this($child)->match(
+                    static fn($found) => $continuation->stop($found),
+                    static fn() => $continuation->continue($found),
+                ),
+            );
+
+        return Maybe::of($found)->keep(
+            Instance::of(Model::class)->or(
+                Instance::of(Custom::class),
+            ),
+        );
     }
 }