+++ tests/Element/BaseTest.php
@@ -6,7 +6,7 @@ namespace Tests\Innmind\Html\Element;
 use Innmind\Html\Element\Base;
 use Innmind\Xml\Element;
 use Innmind\Url\Url;
-use Innmind\Immutable\Set;
+use Innmind\Immutable\Sequence;
 use Innmind\BlackBox\PHPUnit\Framework\TestCase;
 
 class BaseTest extends TestCase
@@ -14,28 +14,31 @@ class BaseTest extends TestCase
     public function testInterface()
     {
         $this->assertInstanceOf(
-            Element::class,
+            Element\Custom::class,
             $base = Base::of(
                 $href = Url::of('http://example.com'),
-                Set::of(),
+                Sequence::of(),
             ),
         );
-        $this->assertSame('base', $base->name());
+        $this->assertSame('base', $base->normalize()->name()->toString());
         $this->assertSame($href, $base->href());
     }
 
     public function testWithoutAttributes()
     {
-        $this->assertTrue(
-            Base::of(Url::of('http://example.com'))->attributes()->empty(),
+        $this->assertFalse(
+            Base::of(Url::of('http://example.com'))->normalize()->attributes()->empty(),
         );
     }
 
     public function testToString()
     {
         $this->assertSame(
-            '<base href="http://example.com/"/>',
-            Base::of(Url::of('http://example.com/'))->toString(),
+            '<base href="http://example.com/"/>'."\n",
+            Base::of(Url::of('http://example.com/'))
+                ->normalize()
+                ->asContent()
+                ->toString(),
         );
     }
 }