+++ tests/Translator/NodeTranslator/ImgTranslatorTest.php
@@ -4,54 +4,21 @@ declare(strict_types = 1);
 namespace Tests\Innmind\Html\Translator\NodeTranslator;
 
 use Innmind\Html\{
-    Translator\NodeTranslator\ImgTranslator,
+    Translator,
     Element\Img,
 };
-use Innmind\Xml\Translator\{
-    Translator,
-    NodeTranslators,
-    NodeTranslator,
-};
+use Innmind\Immutable\Predicate\Instance;
 use Innmind\BlackBox\PHPUnit\Framework\TestCase;
 
 class ImgTranslatorTest extends TestCase
 {
-    public function testInterface()
-    {
-        $this->assertInstanceOf(
-            NodeTranslator::class,
-            ImgTranslator::of(),
-        );
-    }
-
-    public function testReturnNothingWhenNotExpectedElement()
-    {
-        $dom = new \DOMDocument;
-        $dom->loadHTML('<body></body>');
-
-        $result = ImgTranslator::of()(
-            $dom->childNodes->item(1),
-            Translator::of(
-                NodeTranslators::defaults(),
-            )
-        );
-
-        $this->assertNull($result->match(
-            static fn($node) => $node,
-            static fn() => null,
-        ));
-    }
-
     public function testTranslate()
     {
         $dom = new \DOMDocument;
         $dom->loadHTML('<img src="foo.png" alt="bar"/>');
 
-        $img = ImgTranslator::of()(
+        $img = Translator::new()(
             $dom->childNodes->item(1)->childNodes->item(0)->childNodes->item(0),
-            Translator::of(
-                NodeTranslators::defaults(),
-            )
         )->match(
             static fn($img) => $img,
             static fn() => null,
@@ -59,8 +26,9 @@ class ImgTranslatorTest extends TestCase
 
         $this->assertInstanceOf(Img::class, $img);
         $this->assertSame('foo.png', $img->src()->toString());
+        $img = $img->normalize();
         $this->assertCount(2, $img->attributes());
-        $this->assertSame('bar', $img->attributes()->get('alt')->match(
+        $this->assertSame('bar', $img->attribute('alt')->match(
             static fn($attribute) => $attribute->value(),
             static fn() => null,
         ));
@@ -71,12 +39,11 @@ class ImgTranslatorTest extends TestCase
         $dom = new \DOMDocument;
         $dom->loadHTML('<img/>');
 
-        $result = ImgTranslator::of()(
+        $result = Translator::new()(
             $dom->childNodes->item(1)->childNodes->item(0)->childNodes->item(0),
-            Translator::of(
-                NodeTranslators::defaults(),
-            )
-        );
+        )
+            ->maybe()
+            ->keep(Instance::of(Img::class));
 
         $this->assertNull($result->match(
             static fn($node) => $node,