+++ tests/Visitor/HeadTest.php
@@ -8,8 +8,8 @@ use Innmind\Html\{
     Reader\Reader,
 };
 use Innmind\Xml\{
-    Element as ElementInterface,
-    Element\Element,
+    Element,
+    Element\Name,
 };
 use Innmind\Filesystem\File\Content;
 use Innmind\BlackBox\PHPUnit\Framework\TestCase;
@@ -20,7 +20,7 @@ class HeadTest extends TestCase
 
     public function setUp(): void
     {
-        $this->read = Reader::default();
+        $this->read = Reader::new();
     }
 
     public function testExtractHead()
@@ -37,15 +37,15 @@ class HeadTest extends TestCase
             static fn() => null,
         );
 
-        $this->assertInstanceOf(ElementInterface::class, $head);
-        $this->assertSame('head', $head->name());
+        $this->assertInstanceOf(Element::class, $head);
+        $this->assertSame('head', $head->name()->toString());
         $this->assertFalse($head->children()->empty());
         $this->assertTrue($head->attributes()->empty());
     }
 
     public function testReturnNothingWhenHeadNotFound()
     {
-        $this->assertNull(Visitor\Element::head()(Element::of('body'))->match(
+        $this->assertNull(Visitor\Element::head()(Element::of(Name::of('body')))->match(
             static fn($node) => $node,
             static fn() => null,
         ));