+++ CHANGELOG.md
@@ -2,6 +2,27 @@
 
 ## [Unreleased]
 
+### Added
+
+- `Innmind\Html\Translator`
+- `Innmind\Html\Reader\Reader::new()`
+
+### Changed
+
+- Requires `innmind/xml:~8.0`
+- Requires `innmind/filesystem:~8.1`
+- `Innmind\Html\Element\*` classes now implement `Innmind\Xml\Element\Custom`
+- `Innmind\Html\Node\Document` no longer implement any interface
+- `Innmind\Html\Reader\Reader` now return an `Innmind\Immutable\Attempt`
+
+### Removed
+
+- `Innmind\Html\Translator\*`
+- `Innmind\Html\Node\Document::content()`
+- `Innmind\Html\Node\Document::toString()`
+- `Innmind\Html\Reader\Reader::default()`
+- `Innmind\Html\Reader\Reader::of()`
+
 ### Fixed
 
 - PHP `8.4` deprecations